week 4/function calll/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
openai==1.8.5
python-dotenv==1.0.0
pytest==8.4.2
